fix: address PR review feedback across all ports
- go: remove unused vars in trace.go, add context import, fix helpers unused import
- go: nest context under input.context per spec, use math.NaN() on scorer failure
- go: remove local replace directive, use released hop.top/kit v0.3.2-patch.3
- ts: fix scorer field/method collision, use .score() instead of call-as-function
- ts: nest context under input.context per spec
- ts: fix ESM require() in conformance test, remove dead evalYaml
- ts: add ./adapters export to package.json
- php: add toArray() to Trace, Advice, Reward for serialization
- php: fix static method using $this in ConformanceTest
- rs: nest context under input.context per spec
- rs: use f64::NAN on scorer failure
- rs: fix multi-turn state machine (reset state in run(), Done only in multi_turn)
- rs: fix serde_yaml Number conversion
- add .gitignore (vendor, node_modules, dist, target, __pycache__)

+++ rs/src/reward.rs
@@ -97,7 +97,7 @@ impl RewardScorer for CompositeScorer {
         let mut meta = HashMap::new();
         meta.insert(
             "scorers".into(),
-            serde_yaml::Value::Number(rewards.len().into()),
+            serde_yaml::Value::Number((rewards.len() as u64).into()),
         );
 
         Ok(Reward {
